Revert temporarily renaming modules.
