+++ comtypes/_post_coinit/misc.py
@@ -19,7 +19,7 @@ from comtypes import GUID
 from comtypes._idl_stuff import COMMETHOD  # noqa
 from comtypes import CLSCTX_SERVER, CLSCTX_LOCAL_SERVER, CLSCTX_REMOTE_SERVER
 from comtypes import _ole32, oledll, DWORD
-from comtypes._post_coinit._unknwn import IUnknown
+from comtypes._post_coinit.unknwn import IUnknown
 
 
 def _is_object(obj):
